+++ Sources/CFastTextWrapper/include/CFastTextWrapper.hpp
@@ -15,7 +15,7 @@ extern "C" {
     const void * initializeFastTextObject();
     void fasttextLoadModel(const void *object, const char* path);
     int fasttextgetDimension(const void *object);
-    void fasttextgetSentenceVector(const void *object, const char* sentence, double* sentenceVector);
+    void fasttextgetSentenceVector(const void *object, const char* sentence, float* sentenceVector);
 
 #ifdef __cplusplus
 }
